playout/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
PyYAML==6.0.1
watchdog==4.0.0
Jinja2==3.1.3
threefive==2.4.27
